2005-09-20  Dennis Smit  <ds@nerds-incorporated.org>
	* plugins/actor/JESS/*
	* plugins/actor/infinite/*
	* plugins/actor/lv_scope/*:

	Ported all over to new audio API.

	* plugins/actor/oinksie/*: Improved new audio api port.

+++ libvisual-plugins/plugins/actor/infinite/display.h
@@ -10,7 +10,7 @@ void _inf_display (InfinitePrivate *priv, uint8_t *surf);
 void _inf_generate_colors(InfinitePrivate *priv);
 void _inf_change_color(InfinitePrivate *priv, int old_p,int p,int w);
 void _inf_blur(InfinitePrivate *priv, t_interpol* vector_field);
-void _inf_spectral(InfinitePrivate *priv, t_effect* current_effect,short data[2][512]);
+void _inf_spectral(InfinitePrivate *priv, t_effect* current_effect, float data[2][512]);
 void _inf_curve(InfinitePrivate *priv, t_effect* current_effect);
 void _inf_init_display(InfinitePrivate *priv);
